vimrc: added tabline to show buffers
with this commit I added the primitive option to show the open buffers
in the tabline.
this does not work well yet, but it will in time. some more tinkering is
required
I should probably refactor some functions into separate files to make
handling them more efficient, but it works for now.

furthermore trailing whitespaces are now highlighted and some functions
where extracted.